[fix] 마이페이지 내식물보기 패키지수정
